lib/bsd/sys/openbsd/sys/proctable/constants.rb - implement FFI::ConstGenerator for KVM_NO_FILES from kvm.h #tweak

+++ lib/bsd/sys/openbsd/sys/proctable/functions.rb
@@ -6,7 +6,7 @@ module Sys
     ffi_lib :kvm
 
     # attach_function :devname, [:dev_t, :mode_t], :string
-    attach_function :kvm_openfiles, [:string, :string, :string, :long, :string], :pointer
+    attach_function :kvm_openfiles, [:string, :string, :string, :int, :string], :pointer
     attach_function :kvm_close, [:pointer], :int
     attach_function :kvm_getprocs, [:pointer, :int, :int, :size_t, :pointer], :pointer
     attach_function :kvm_getargv, [:pointer, :pointer, :int], :pointer
